test(facilitator): extract shared verify+settle fixtures
CodeRabbit flagged that the `fullOffer` literal is duplicated between
verify.test.ts and settle.test.ts and will silently drift the next
time the Boson `FullOffer` struct changes. Per the repo's own
CLAUDE.md rule ("move duplicated constants to a shared file the
moment they appear in two places"), the same logic applies to all the
other scalars and helpers both files copy:

- BUYER_PK / RELAYER_PK + the two viem accounts
- ESCROW / ASSET / CHAIN_ID / NETWORK / NONCE / SELLER / SELLER_SIG / AMOUNT
- the canonical fullOffer object
- buildValidPayload() + buildValidRequirements()

Extracted into test/fixtures.ts; both test files now import from there.
Per-file mock builders (buildPublicClient differs between suites
because settle additionally needs waitForTransactionReceipt;
buildWalletClient + buildReceipt are settle-only) stay local —
those differences are part of each test's intent.

verify.test.ts and settle.test.ts each shrink by ~145 lines. Test
count unchanged at 40 passing; the consolidation is pure dedup with
no behavioral change.
